+++ src/main/java/org/jabref/logic/integrity/PagesChecker.java
@@ -1,5 +1,6 @@
 package org.jabref.logic.integrity;
 
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
@@ -11,28 +12,26 @@ import org.jabref.model.strings.StringUtil;
 public class PagesChecker implements ValueChecker {
 
     private static final String PAGES_EXP_BIBTEX = ""
-            + "\\A"                 // begin String
+            + "\\A"               // begin String
             + "("
             + "[A-Za-z]?\\d*"     // optional prefix and number
             + "("
-            + "\\+|-{2}"    // separator
-            + "[A-Za-z]?\\d*" // optional prefix and number
+            + "\\+|-{2}"          // separator
+            + "[A-Za-z]?\\d*"     // optional prefix and number
             + ")?"
             + ",?"                // page range separation
             + ")*"
-            + "\\z";                // end String
+            + "\\z";              // end String
 
+    // See https://packages.oth-regensburg.de/ctan/macros/latex/contrib/biblatex/doc/biblatex.pdf#subsubsection.3.15.3 for valid content
     private static final String PAGES_EXP_BIBLATEX = ""
-            + "\\A"                 // begin String
-            + "("
+            + "\\A"               // begin String
             + "[A-Za-z]?\\d*"     // optional prefix and number
             + "("
-            + "\\+|-{1,2}|\u2013" // separator
-            + "[A-Za-z]?\\d*" // optional prefix and number
+            + "(\\+|-{1,2}|\u2013)" // separator
+            + "[A-Za-z]?\\d*"     // optional prefix and number
             + ")?"
-            + ",?"                // page range separation
-            + ")*"
-            + "\\z";                // end String
+            + "\\z";              // end String
 
     private final Predicate<String> isValidPageNumber;
 
@@ -59,10 +58,13 @@ public class PagesChecker implements ValueChecker {
             return Optional.empty();
         }
 
-        if (!isValidPageNumber.test(value.trim())) {
+        String[] split = value.split(",");
+
+        if (Arrays.stream(split)
+                .map(String::trim)
+                .anyMatch(pageRange -> !isValidPageNumber.test(pageRange))) {
             return Optional.of(Localization.lang("should contain a valid page number range"));
         }
-
         return Optional.empty();
     }
 }